+++ src/core_modules/capture-core/components/ScopeSelector/hooks/useResetStageId.js
@@ -1,15 +1,17 @@
 // @flow
 import { useHistory, useLocation } from 'react-router-dom';
-import { getUrlQueries } from '../../../utils/url';
-import { buildUrlQueryString } from '../../../utils/routing';
+import { buildUrlQueryString, useLocationQuery } from '../../../utils/routing';
+import type { PageContext } from './types';
 
 export const useResetStageId = () => {
     const history = useHistory();
     const { pathname } = useLocation();
+    const { programId, orgUnitId, teiId, enrollmentId } = useLocationQuery();
 
-    const resetStageId = (pageToPush: string = pathname) => {
-        const { programId, orgUnitId, teiId, enrollmentId } = getUrlQueries();
-        history.push(`${pageToPush}?${buildUrlQueryString({ programId, orgUnitId, teiId, enrollmentId })}`);
+    const resetStageId = (pageToPush: string = pathname, pageContext?: PageContext) => {
+        history.push(`${pageToPush}?${buildUrlQueryString({
+            programId, orgUnitId, teiId, enrollmentId: enrollmentId ?? pageContext?.enrollmentId,
+        })}`);
     };
 
     return { resetStageId };