+++ src/core_modules/capture-core/components/TopBarActions/index.js
@@ -1,3 +1,2 @@
 // @flow
-export { topBarActionsActionTypes } from './TopBarActions.actions';
 export { TopBarActions } from './TopBarActions.container';