+++ src/core_modules/capture-core/reducers/descriptions/app.reducerDescriptionGetter.js
@@ -9,19 +9,18 @@ import { actionTypes as editEventActionTypes }
     from '../../components/Pages/ViewEvent/ViewEventComponent/editEvent.actions';
 import { actionTypes as viewEventActionTypes }
     from '../../components/Pages/ViewEvent/ViewEventComponent/viewEvent.actions';
-import {
-    actionTypes as editEventDataEntryActionTypes,
-} from '../../components/WidgetEventEdit/DataEntry/editEventDataEntry.actions';
+import { newPageActionTypes }
+    from '../../components/Pages/New/NewPage.actions';
 import { actionTypes as connectivityActionTypes } from '../../components/Connectivity/connectivity.actions';
 import {
     actionTypes as setCurrentSelectionsActionTypes,
-} from '../../components/LockedSelector/QuickSelector/actions/QuickSelector.actions';
+} from '../../components/ScopeSelector/QuickSelector/actions/QuickSelector.actions';
 import {
     lockedSelectorActionTypes,
 } from '../../components/LockedSelector';
-import { topBarActionsActionTypes } from '../../components/TopBarActions';
 import { searchPageActionTypes } from '../../components/Pages/Search/SearchPage.actions';
 import { enrollmentPageActionTypes } from '../../components/Pages/Enrollment/EnrollmentPage.actions';
+import { actionTypes as initActionTypes } from '../../init/init.actions';
 
 const LOCATION_CHANGE = '@@router/LOCATION_CHANGE';
 const OFFLINE_STATUS_CHANGED = 'Offline/STATUS_CHANGED';
@@ -43,18 +42,6 @@ export const getAppReducerDesc = (appUpdaters: Updaters) => createReducerDescrip
         return newState;
     },
     [newEventDataEntryActionTypes.START_CANCEL_SAVE_RETURN_TO_MAIN_PAGE]: (state) => {
-        const newState = { ...state };
-        newState.page = null;
-        newState.locationSwitchInProgress = true;
-        return newState;
-    },
-    [editEventDataEntryActionTypes.REQUEST_SAVE_RETURN_TO_MAIN_PAGE]: (state) => {
-        const newState = { ...state };
-        newState.page = null;
-        newState.locationSwitchInProgress = true;
-        return newState;
-    },
-    [editEventDataEntryActionTypes.START_CANCEL_SAVE_RETURN_TO_MAIN_PAGE]: (state) => {
         const newState = { ...state };
         newState.page = null;
         newState.locationSwitchInProgress = true;
@@ -122,46 +109,27 @@ export const getAppReducerDesc = (appUpdaters: Updaters) => createReducerDescrip
         };
         return newState;
     },
-
-    [lockedSelectorActionTypes.ORG_UNIT_ID_SET]: state => ({
-        ...state,
-        locationSwitchInProgress: true,
-    }),
-    [lockedSelectorActionTypes.PROGRAM_ID_SET]: state => ({
-        ...state,
-        locationSwitchInProgress: true,
-    }),
     [lockedSelectorActionTypes.FROM_URL_UPDATE]: (state, action) => ({
         ...state,
         page: action.payload.nextPage,
     }),
-    [lockedSelectorActionTypes.NEW_REGISTRATION_PAGE_OPEN]: state => ({
+    [newPageActionTypes.NEW_PAGE_OPEN]: state => ({
         ...state,
+        locationSwitchInProgress: false,
         page: 'new',
-    }),
-    [topBarActionsActionTypes.NEW_REGISTRATION_PAGE_OPEN]: state => ({
-        ...state,
-        page: 'new',
-    }),
-    [lockedSelectorActionTypes.ORG_UNIT_ID_RESET]: state => ({
-        ...state,
-        locationSwitchInProgress: true,
-    }),
-    [lockedSelectorActionTypes.PROGRAM_ID_RESET]: state => ({
-        ...state,
-        locationSwitchInProgress: true,
     }),
     [searchPageActionTypes.TO_MAIN_PAGE_NAVIGATE]: state => ({
         ...state,
         page: null,
         locationSwitchInProgress: true,
     }),
-    [lockedSelectorActionTypes.SEARCH_PAGE_OPEN]: state => ({
+    [searchPageActionTypes.SEARCH_PAGE_OPEN]: state => ({
         ...state,
+        locationSwitchInProgress: false,
         page: 'search',
     }),
-    [topBarActionsActionTypes.SEARCH_PAGE_OPEN]: state => ({
+    [initActionTypes.SET_CURRENT_ORG_UNIT_ROOT]: (state, action) => ({
         ...state,
-        page: 'search',
+        roots: action.payload.roots,
     }),
 }, 'app');