+++ src/core_modules/capture-core/storageControllers/userStorageController.js
@@ -1,10 +1,8 @@
 // @flow
 import { StorageController, IndexedDBAdapter } from 'capture-core-utils/storage';
-import { getCurrentUser } from '../d2/d2Instance';
 import { userStores } from './stores/index';
 
-function getStorageName(mainStorageName: string) {
-    const user = getCurrentUser();
+function getStorageName(mainStorageName: string, user: { id: string }) {
     return mainStorageName + user.id;
 }
 
@@ -15,8 +13,9 @@ function getStores() {
 
 function createStorageController(
     mainStorageController: typeof StorageController,
+    currentUser: { id: string },
 ) {
-    const storageName = getStorageName(mainStorageController.name);
+    const storageName = getStorageName(mainStorageController.name, currentUser);
     const appCacheVersion = mainStorageController.version;
     const stores = getStores();
     const storageController =
@@ -49,9 +48,9 @@ const storeSpecificCreateActions = {
     },
 };
 
-export async function initUserControllerAsync(mainStorageController: typeof StorageController) {
+export async function initUserControllerAsync(mainStorageController: typeof StorageController, currentUser: { id: string }) {
     const userStorageController =
-        createStorageController(mainStorageController);
+        createStorageController(mainStorageController, currentUser);
 
     let upgradeTempData;
     await userStorageController