+++ src/core_modules/capture-core/utils/url.js
@@ -1,6 +1,6 @@
 // @flow
 import { pageKeys } from '../components/App/withAppUrlSync';
-import { deriveURLParamsFromLocation } from './routing';
+import { getLocationQuery } from './routing';
 
 export type Url = {
     programId?: string,
@@ -29,7 +29,7 @@ export const deriveUrlQueries = (state: Object) => {
         trackedEntityTypeId: routerTet,
         teiId,
         enrollmentId,
-    } = deriveURLParamsFromLocation();
+    } = getLocationQuery();
 
     const programId = routerProgramId || selectedProgramId;
     const orgUnitId = routerOrgUnitId || selectedOrgUnitId;
@@ -56,16 +56,6 @@ export const getLocationPathname = () => {
 // TODO - This will be removed when the link to tracker capture is removed
 export const getLocationSearch = () => window.location.search;
 
-export const getUrlQueries = (): Url => {
-    const split = window.location.href.split('?');
-    const searchParams = new URLSearchParams(split && split[1]);
-    let searchParamsObject: Url = {};
-
-    for (const [key, value] of searchParams.entries()) {
-        searchParamsObject = { ...searchParamsObject, [key]: value };
-    }
-    return searchParamsObject;
-};
 
 export const pageFetchesOrgUnitUsingTheOldWay = (page: string, pages: Object = pageKeys): boolean =>
     Object.values(pages).includes(page);